- When opening the DifferencesPane select base and compare versions to represent the   differences in the currently selected versions in the history view.  Fixes Issue 38.

+++ Classes/General/Common.h
@@ -229,8 +229,8 @@ typedef enum
 
 typedef struct
 {
-	int	lowRevision;
-	int	highRevision;
+	NSInteger	lowRevision;
+	NSInteger	highRevision;
 } LowHighPair;
 
 typedef struct
